Admin Games fix for handicaps
We needed an upsert  flag of true when writing the handicaps so when the first one of the season is made it gets into the db.  Only one document per season contains the handicaps as of each date.
Fixed mongo.js so it will use the database in the uri we supply we don't need to hardcode the db .... so local host can use any db you want